+++ src/veilwright/utility.py
@@ -1,13 +1,13 @@
 """Utility Judge for scoring meaning preservation."""
 
-from defender.llm_client import GeminiClient
-from defender.prompts import (
+from veilwright.llm_client import GeminiClient
+from veilwright.prompts import (
     UTILITY_RETRY_PROMPT,
     UTILITY_SYSTEM_PROMPT,
     build_utility_prompt,
 )
-from defender.types import UtilityInput, UtilityOutput
-from defender.utils import parse_validated_llm_json, validate_utility_response
+from veilwright.types import UtilityInput, UtilityOutput
+from veilwright.utils import parse_validated_llm_json, validate_utility_response
 
 
 class UtilityError(Exception):